# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/category/create": {
      "post": {
        "operationId": "CategoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建分类",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/list": {
      "get": {
        "operationId": "CategoryController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码，从 1 开始",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数，默认 20，最大 100",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "moduleCode",
            "required": false,
            "in": "query",
            "description": "模块编码",
            "schema": {
              "example": "article",
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "required": false,
            "in": "query",
            "description": "父分类 ID（传 0 或 null 查询根分类）",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态",
            "schema": {
              "type": "string",
              "enum": [
                "enabled",
                "disabled"
              ]
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字搜索（名称）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "是否包含已删除",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "default": "sortOrder",
              "type": "string",
              "enum": [
                "id",
                "createdAt",
                "updatedAt",
                "sortOrder",
                "name"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "分页查询分类列表",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/stats": {
      "get": {
        "operationId": "CategoryController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取分类统计数据",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/tree/{moduleCode}": {
      "get": {
        "operationId": "CategoryController_getTree",
        "parameters": [
          {
            "name": "moduleCode",
            "required": true,
            "in": "path",
            "description": "模块编码",
            "schema": {
              "example": "article",
              "type": "string"
            }
          },
          {
            "name": "onlyEnabled",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取指定模块的分类树",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/options/{moduleCode}": {
      "get": {
        "operationId": "CategoryController_getOptions",
        "parameters": [
          {
            "name": "moduleCode",
            "required": true,
            "in": "path",
            "description": "模块编码",
            "schema": {
              "example": "article",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取分类选项列表（用于下拉选择）",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/slug/{moduleCode}/{slug}": {
      "get": {
        "operationId": "CategoryController_findBySlug",
        "parameters": [
          {
            "name": "moduleCode",
            "required": true,
            "in": "path",
            "description": "模块编码",
            "schema": {
              "example": "article",
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "URL 标识",
            "schema": {
              "example": "tech-articles",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据模块和 slug 查询分类",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/uid/{uid}": {
      "get": {
        "operationId": "CategoryController_findByUid",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "description": "分类 UID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据 UID 查询分类详情",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/detail/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据 ID 查询分类详情",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/update/{id}": {
      "post": {
        "operationId": "CategoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新分类",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/enable/{id}": {
      "post": {
        "operationId": "CategoryController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "启用分类",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/disable/{id}": {
      "post": {
        "operationId": "CategoryController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "禁用分类",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/move/{id}": {
      "post": {
        "operationId": "CategoryController_move",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "parentId": {
                    "type": "number",
                    "nullable": true,
                    "description": "新的父分类 ID，null 表示移动到根级"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "移动分类（更改父分类）",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/restore/{id}": {
      "post": {
        "operationId": "CategoryController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "恢复已删除的分类",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/delete/{id}": {
      "post": {
        "operationId": "CategoryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "分类 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除分类（软删除）",
        "tags": [
          "分类管理"
        ]
      }
    },
    "/admin/category/batch-delete": {
      "post": {
        "operationId": "CategoryController_batchRemove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "批量删除分类（只删除叶子节点）",
        "tags": [
          "分类管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "jmni-server",
    "description": "JMNI Server API 文档 - 供外部团队参考",
    "version": "0.0.1",
    "contact": {
      "name": "JMNI Team",
      "url": "https://jmni.cn",
      "email": "support@jmni.cn"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://api.jmni.cn",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:2233",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT Access Token"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "moduleCode": {
            "type": "string",
            "description": "模块编码",
            "example": "article",
            "maxLength": 64
          },
          "name": {
            "type": "string",
            "description": "分类名称",
            "example": "技术文章",
            "maxLength": 64
          },
          "slug": {
            "type": "string",
            "description": "URL 标识（同一模块内唯一）",
            "example": "tech-articles",
            "maxLength": 128
          },
          "description": {
            "type": "string",
            "description": "分类描述",
            "maxLength": 255
          },
          "icon": {
            "type": "string",
            "description": "图标",
            "maxLength": 128
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图 URL",
            "maxLength": 512
          },
          "parentId": {
            "type": "number",
            "description": "父分类 ID",
            "example": 1
          },
          "sortOrder": {
            "type": "number",
            "description": "排序权重（越大越靠前）",
            "example": 0,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "状态",
            "enum": [
              "enabled",
              "disabled"
            ],
            "default": "enabled"
          }
        },
        "required": [
          "moduleCode",
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名称",
            "maxLength": 64
          },
          "slug": {
            "type": "string",
            "description": "URL 标识",
            "maxLength": 128
          },
          "description": {
            "type": "string",
            "description": "分类描述",
            "maxLength": 255
          },
          "icon": {
            "type": "string",
            "description": "图标",
            "maxLength": 128
          },
          "coverUrl": {
            "type": "string",
            "description": "封面图 URL",
            "maxLength": 512
          },
          "parentId": {
            "type": "number",
            "description": "父分类 ID"
          },
          "sortOrder": {
            "type": "number",
            "description": "排序权重",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "状态",
            "enum": [
              "enabled",
              "disabled"
            ]
          }
        }
      },
...
    }
  }
}